// src/defs/enums.ts
export enum Planet {
    Sun = "Sun",
    Moon = "Moon",
    Mercury = "Mercury",
    Venus = "Venus",
    Mars = "Mars",
    Jupiter = "Jupiter",
    Saturn = "Saturn",
    Uranus = "Uranus",
    Neptune = "Neptune",
    Pluto = "Pluto",
    Chiron = "Chiron",
    Ascendant = "Ascendant",
    TrueNorthNode = "True North Node",
    TrueSouthNode = "True South Node",
    Lilith = "Lilith",
    Nadir = "Nadir",
    Descendant = "Descendant",
    MidHeaven = "MidHeaven",
}

export enum PlanetId {
    Sun = 0,
    Moon = 1,
    Mercury = 2,
    Venus = 3,
    Mars = 4,
    Jupiter = 5,
    Saturn = 6,
    Uranus = 7,
    Neptune = 8,
    Pluto = 9,
    Chiron = 15,
    Ascendant = 100,
    TrueNorthNode = 103,
    TrueSouthNode = 104,
    Lilith = 105,
    Nadir = 107,
    Descendant = 108,
    MidHeaven = 109,
}

export enum ZodiacSign {
    Aries = "Aries",
    Taurus = "Taurus",
    Gemini = "Gemini",
    Cancer = "Cancer",
    Leo = "Leo",
    Virgo = "Virgo",
    Libra = "Libra",
    Scorpio = "Scorpio",
    Sagittarius = "Sagittarius",
    Capricorn = "Capricorn",
    Aquarius = "Aquarius",
    Pisces = "Pisces",
}

export enum Aspect {
    Conjunction = "Conjunction",
    Opposition = "Opposition",
    Square = "Square",
    SemiSquare = "Semi Square",
    Sesquiquadrate = "Sesquiquadrate",
    Trine = "Trine",
    Sextile = "Sextile",
    SemiSextile = "Semi Sextile",
    Quincunx = "Quincunx",
    Quintile = "Quintile",
    BiQuintile = "Bi Quintile",
    Parallel = "Parallel",
    Contraparallel = "Contraparallel",
}